be/requirements.txt
django==3.0.7
djangorestframework==3.10.0
djangorestframework-simplejwt==4.6.0
PyJWT==2.0.0
django-cors-headers==3.4.0
djoser==2.1.0
Pillow==9.5.0
